Add files via upload
Fully functioning register, login, logout pages. The header changes depending on if your are logged in or logged out (if you are logged in the login and register pages are hidden and logout is displayed, and if you are not logged in then the logout page is hidden and login and register pages are displayed). Create, Edit and Delete functions are hidden if you are not logged in yet.